etls/requirements.txt
alembic==1.4.2
astroid==2.4.2
backoff==1.10.0
certifi==2020.6.20
chardet==3.0.4
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
numpy==1.19.1
pandas==1.1.1
pangres==2.2
psycopg2==2.8.5
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
ratelimit==2.2.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
tweet-preprocessor==0.6.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
wrapt==1.12.1
yapf==0.30.0
